+++ Sources/VCore/Models/Layout Directions/LayoutDirectionOmni.swift
@@ -56,7 +56,7 @@ public enum LayoutDirectionOmni: CaseIterable {
     ///
     /// For `leftToRight` or `rightToLeft`, `HStack` will be used.
     /// For `topToBottom` or `bottomToTop`, `VStack` will be used.
-    @ViewBuilder public func stackView<Content>(
+    public func stackView<Content>(
         alignmentHor: HorizontalAlignment = .center,
         alignmentVer: VerticalAlignment = .center,
         spacing: CGFloat? = nil,
@@ -64,20 +64,13 @@ public enum LayoutDirectionOmni: CaseIterable {
     ) -> some View
         where Content: View
     {
-        if isHorizontal {
-            HStack(
-                alignment: alignmentVer,
-                spacing: spacing,
-                content: content
-            )
-            
-        } else {
-            VStack(
-                alignment: alignmentHor,
-                spacing: spacing,
-                content: content
-            )
-        }
+        HOrVStack(
+            alignmentHor: alignmentHor,
+            alignmentVer: alignmentVer,
+            spacing: spacing,
+            isHorizontal: isHorizontal,
+            content: content
+        )
     }
     
     /// Creates stack `Layout` based on axis of direction.